moved + renamed (inference) model api
