+++ frontend/lost/src/actions/pipeline/model/pipeline-template-response.ts
@@ -1,4 +1,4 @@
-import { Model } from '../../inference-model/model-api'
+import { Model } from '../../../api/inference_model'
 
 export interface PipelineTemplateResponse {
   id: number
